requirements.txt
amqp==5.3.1
anyio==3.7.1
apischema
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
bcrypt==4.3.0
billiard==4.2.1
cachetools==4.2.4
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
construct==2.10.68
construct-typing==0.5.6
cryptography==44.0.2
Django==4.2.20
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
exceptiongroup==1.2.2
flake8==7.2.0
h11==0.12.0
httpcore
httpx
idna==3.10
Jinja2==3.1.6
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
MarkupSafe==3.0.2
mccabe==0.7.0
OSlash==0.6.3
paramiko==3.5.1
pillow==11.2.1
prompt_toolkit==3.0.51
publickey==0.3
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.0
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
solana==0.25.0
sqlparse==0.5.3
types-cachetools==4.2.10
typing_extensions
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==10.4
djangorestframework-simplejwt~=5.5.0